reed_solomon_simd::engine

Struct ShardsRefMut

pub struct ShardsRefMut<'a> { /* private fields */ }

Mutable reference to a shard array.

Implementations

impl<'a> ShardsRefMut<'a>

pub fn dist2_mut( &mut self, pos: usize, dist: usize, ) -> (&mut [[u8; 64]], &mut [[u8; 64]])

Returns mutable references to shards at pos and pos + dist.

See source code of Naive::fft for an example.

Panics

If dist is 0.

pub fn dist4_mut( &mut self, pos: usize, dist: usize, ) -> (&mut [[u8; 64]], &mut [[u8; 64]], &mut [[u8; 64]], &mut [[u8; 64]])

Returns mutable references to shards at pos, pos + dist, pos + dist * 2 and pos + dist * 3.

See source code of NoSimd::fft for an example (specifically the private method fft_butterfly_two_layers).

Panics

If dist is 0.

pub fn is_empty(&self) -> bool

Returns true if this contains no shards.

pub fn len(&self) -> usize

Returns number of shards.

pub fn new( shard_count: usize, shard_len_64: usize, data: &'a mut [[u8; 64]], ) -> Self

Creates new ShardsRefMut that references given data.

Panics

If data.len() < shard_count * shard_len_64.

pub fn split_at_mut( &mut self, mid: usize, ) -> (ShardsRefMut<'_>, ShardsRefMut<'_>)

Splits this ShardsRefMut into two so that first includes shards 0..mid and second includes shards mid...

pub fn zero<R: RangeBounds<usize>>(&mut self, range: R)

Fills the given shard-range with 0u8:s.