reed_solomon_simd::engine

Module tables

Lookup-tables used by Engine:s.

All tables are global and each is initialized at most once.

Tables

Table	Size	Used in encoding	Used in decoding	By engines

Exp	128 kiB	yes	yes	all

Log	128 kiB	yes	yes	all

LogWalsh	128 kiB	-	yes	all

Mul16	8 MiB	yes	yes	NoSimd

Mul128	8 MiB	yes	yes	Avx2 Ssse3

Skew	128 kiB	yes	yes	all

Structs

ExpLog

Struct holding the Exp and Log lookup tables.

Multiply128lutT

Elements of the Mul128 table

Functions

get_exp_log

Lazily initialized exponentiation and logarithm tables.

get_log_walsh

Lazily initialized logarithmic Walsh transform table.

get_mul16

Lazily initialized multiplication table for the NoSimd engine.

get_mul128

Lazily initialized multiplication table for SIMD engines.

get_skew

Lazily initialized skew table used in FFT and IFFT operations.

mul

Calculates x * log_m using Exp and Log tables.

Type Aliases

Exp

Used by Naive engine for multiplications and by all Engine:s to initialize other tables.

Log

Used by Naive engine for multiplications and by all Engine:s to initialize other tables.

LogWalsh

Used by all Engine:s in Engine::eval_poly.

Mul16

Used by NoSimd engine for multiplications.

Mul128

Used by Avx2 and Ssse3 engines for multiplications.

Skew

Used by all Engine:s for FFT and IFFT.