reed_solomon_simd

Module rate

Advanced encoding/decoding using chosen Engine and Rate.

This is an advanced module which is not needed for simple usage or basic usage.

This module is relevant if you want to

encode/decode using other Engine than DefaultEngine.

re-use working space of one encoder/decoder in another.

understand/benchmark/test high or low rate directly.

Rates

See algorithm > Rate for details about high/low rate.

DefaultRate, DefaultRateEncoder, DefaultRateDecoder

Encoding/decoding using high or low rate as appropriate.

These are basically same as ReedSolomonEncoder and ReedSolomonDecoder except with slightly different API which allows specifying Engine and working space.

HighRate, HighRateEncoder, HighRateDecoder

Encoding/decoding using only high rate.

LowRate, LowRateEncoder, LowRateDecoder

Encoding/decoding using only low rate.

Structs

DecoderWork

Working space for RateDecoder.

DefaultRate

Reed-Solomon encoder/decoder generator using high or low rate as appropriate.

DefaultRateDecoder

Reed-Solomon decoder using high or low rate as appropriate.

DefaultRateEncoder

Reed-Solomon encoder using high or low rate as appropriate.

EncoderWork

Working space for RateEncoder.

HighRate

Reed-Solomon encoder/decoder generator using only high rate.

HighRateDecoder

Reed-Solomon decoder using only high rate.

HighRateEncoder

Reed-Solomon encoder using only high rate.

LowRate

Reed-Solomon encoder/decoder generator using only low rate.

LowRateDecoder

Reed-Solomon decoder using only low rate.

LowRateEncoder

Reed-Solomon encoder using only low rate.

Traits

Rate

Reed-Solomon encoder/decoder generator using specific rate.

RateDecoder

Reed-Solomon decoder using specific rate.

RateEncoder

Reed-Solomon encoder using specific rate.